+++ Lemonade Stand/Assets/Scripts/IState.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+public interface IState
+{
+    void Enter(); // runs when you enter the state
+    void Exit(); // runs when you exit the state
+    void Tick(); // runs every update tick
+}