+++ Lemonade Stand/Assets/Scripts/PrepDayState.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+
+public class PrepDayState : GameState
+{
+    public PrepDayState(GameManager gameManager) : base(gameManager){ }
+
+    public override void Enter()
+    {
+        
+    }
+
+    public override void Exit()
+    {
+        
+    }
+
+    public override void Tick()
+    {
+        
+    }
+}