+++ Lemonade Stand/Assets/Scripts/StateMachine.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+public class StateMachine
+{
+    private IState CurrentState { get; set; }
+
+    public void ChangeState(IState newState)
+    {
+        CurrentState?.Exit();
+        CurrentState = newState;
+        CurrentState.Enter();
+    }
+
+    public void Tick()
+    {
+        CurrentState?.Tick();
+    }
+}